confirm if user wants to update password

+++ frontend/src/screens/ProfileEditScreen.js
@@ -30,7 +30,6 @@ export default function ProfileEditScreen() {
   const [name, setName] = useState(userInfo.name);
   const [email, setEmail] = useState(userInfo.email);
   const [password, setPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
   const [confirmPasswordToSeeUserInfo, setConfirmPasswordToSeeUserInfo] =
     useState(false);
   const [disabledValue, setDisabledValue] = useState(true);
@@ -43,29 +42,31 @@ export default function ProfileEditScreen() {
 
   const submitHandler = async (e) => {
     e.preventDefault();
-    try {
-      const { data } = await axios.put(
-        '/api/users/profile/edit',
-        {
-          name,
-          email,
-          password,
-        },
-        {
-          headers: { Authorization: `Bearer ${userInfo.token}` },
-        }
-      );
-      dispatch({
-        type: 'UPDATE_SUCCESS',
-      });
-      ctxDispatch({ type: 'USER_SIGNIN', payload: data });
-      localStorage.setItem('userInfo', JSON.stringify(data));
-      toast.success('User updated successfully');
-    } catch (err) {
-      dispatch({
-        type: 'FETCH_FAIL',
-      });
-      toast.error(getError(err));
+    if (window.confirm('Are you sure you want to update your username?')) {
+      try {
+        const { data } = await axios.put(
+          '/api/users/profile/edit',
+          {
+            name,
+            email,
+            password,
+          },
+          {
+            headers: { Authorization: `Bearer ${userInfo.token}` },
+          }
+        );
+        dispatch({
+          type: 'UPDATE_SUCCESS',
+        });
+        ctxDispatch({ type: 'USER_SIGNIN', payload: data });
+        localStorage.setItem('userInfo', JSON.stringify(data));
+        toast.success('User updated successfully');
+      } catch (err) {
+        dispatch({
+          type: 'FETCH_FAIL',
+        });
+        toast.error(getError(err));
+      }
     }
   };
